refactor(cli): Standardizes name of command files Switches from dot-delimited  to kebab case for init and compile commands
