+++ cli/src/commands/compile/compile-service.ts
